Removed pfp_rpNN.py.
